tja: add shift jis support
This commit adds support for Shift JIS in the TJA frontend, where
character conversion routines (namely the iconv set of functions)
are available on the system. libtaco expects all strings to be encoded
as UTF-8.

Currently, simdutf, a fast C++ library for Unicode validation, is
required as a hard dependency of Shift JIS support. This may be relaxed
into a soft dependency in the future.